refactor: colocate raw input tests
